@@ -1,50 +1,51 @@
-source 'https://rubygems.org'
-ruby '2.2.0'
+source "https://rubygems.org"
+ruby "2.2.0"
 
-gem 'rails', '4.1.0'
+gem "rails", "4.2.0"
 
-gem 'sass-rails', '~> 4.0.0'
-gem 'coffee-rails', '~> 4.0.0'
-gem 'jquery-rails', '~> 2.2.1'
-gem 'jquery-tablesorter', '1.9.5'
-gem 'bootstrap-sass', '~> 3.1.1'
-gem 'uglifier', '>= 1.3.0'
-gem 'slim-rails', '~> 2.1.0'
-gem 'pg', '0.17.1'
-gem 'jbuilder', '~> 1.2'
-gem 'google-analytics-rails', '~> 0.0.5'
-gem 'devise', '~> 3.2.3'
-gem 'cancancan', '~> 1.7'
-gem 'newrelic_rpm'
-gem 'simple_form', '~> 3.0.2'
-gem 'omniauth-github', :git => 'git://github.com/intridea/omniauth-github.git'
-gem 'friendly_id', '~> 5.0.3'
-gem 'turbolinks', '~> 2.3.0'
-gem 'jquery-turbolinks', '~> 2.1.0'
+gem "sass-rails", "~> 4.0.0"
+gem "coffee-rails", "~> 4.0.0"
+gem "jquery-rails", "~> 2.2.1"
+gem "jquery-tablesorter", "1.9.5"
+gem "bootstrap-sass", "~> 3.1.1"
+gem "uglifier", ">= 1.3.0"
+gem "slim-rails", "~> 3.0.1"
+gem "pg", "0.17.1"
+gem "jbuilder", "~> 1.2"
+gem "google-analytics-rails", "~> 0.0.5"
+gem "devise", "~> 3.4.0"
+gem "cancancan", "~> 1.7"
+gem "newrelic_rpm"
+gem "simple_form", "~> 3.0.2"
+gem "omniauth-github", "1.1.2"
+gem "friendly_id", "~> 5.0.3"
+gem "turbolinks", "~> 2.3.0"
+gem "jquery-turbolinks", "~> 2.1.0"
 
-gem 'pry-rails'
-gem 'awesome_print'
+gem "pry-rails"
+gem "awesome_print"
 
 # For Lectures
 gem "rails-reveal-js", "~> 2.6.1.2"
 
 group :doc do
-  gem 'sdoc', require: false
+  gem "sdoc", require: false
 end
 
 group :development, :test do
-  gem 'annotate', '~> 2.6.1'
-  gem 'better_errors', '~> 1.1.0'
-  gem 'binding_of_caller', '~> 0.7.2'
-  gem 'bullet', '~> 4.8.0'
-  gem 'commands', '~> 0.2.1'
-  gem 'factory_girl', '~> 4.4.0'
-  gem 'rails_best_practices', '~> 1.15.1'
-  gem 'rspec-rails', '~> 2.14.1'
-  gem 'rails-rename', '~> 1.0.0'
-  gem 'quiet_assets', '~> 1.0.2'
+  gem "annotate", "~> 2.6.1"
+  gem "better_errors", "~> 1.1.0"
+  gem "binding_of_caller", "~> 0.7.2"
+  gem "bullet", "~> 4.8.0"
+  gem "commands", "~> 0.2.1"
+  gem "factory_girl_rails", "~> 4.4.0"
+  gem "figaro", "1.1.0"
+  gem "rails_best_practices", "~> 1.15.1"
+  gem "rspec-rails", "~> 2.14.1"
+  gem "rails-rename", "~> 1.0.0"
+  gem "quiet_assets", "~> 1.0.2"
 end
 
 group :production do
-  gem 'rails_12factor'
+  gem "rails_12factor"
 end
